==> 2. convert_raw_pickel_data.ipynb <==
import pickle
import pandas as pd
from tqdm import tqdm

### Load pickle data

with open('data/raw_data_mod21c3.pkl', 'rb') as f:
    data = pickle.load(f)

list = []
coordinate = [str(item[1:3]) for item in data[0]]

for day in tqdm(range(len(data))):
    daily_data = data[day]
    date = daily_data[1][0]
    year = date.split("_")[0]
    temp_list = [item[-1] for item in daily_data]
    temp_list.insert(0, year)
    temp_list.insert(1, date)
    list.append(temp_list)
    
df = pd.DataFrame(list, columns=['year', 'date'] + coordinate)

df.head()

# TODO: remove label columns 
df.drop(df.columns[2], axis=1, inplace=True)

df.head()

df.to_csv('data/data_mod21c3.csv', index=False)
