==> 3. split_data.ipynb <==
import pandas as pd

df = pd.read_csv('data/data_mod21c3.csv')

df.head(15)

df.describe()

# Count null values
df.isna().sum().sum()

### Drop columns with all null values

# Count columns with all null values
df.isna().all().sum()

# Drop columns with all null values
df = df.dropna(axis=1, how='all')

# Check if columns with all null values are dropped
df.head(15) 

### Fill null values "in between" with linear interpolation

# Fill null values "in between" with linear interpolation
df = df.interpolate(method='linear') 

df.isna().sum().sum()

# Remove columns when there are still null values
# df = df.dropna(axis=1, how='any')

### Fill null values on the first few row with the next value

# Fill null value on the first few row with the next value
df = df.bfill()

### Make sure no null values

df.describe()

df.isna().sum().sum()

df.head(15)

print('Minimum Values:', df.drop(['year', 'date'], axis=1).min().min())
print('Maximum Values:', df.drop(['year', 'date'], axis=1).max().max())
print('Total Count:', df.drop(['year', 'date'], axis=1).shape[0] * df.drop(['year', 'date'], axis=1).shape[1])

### Split data into training-validation, test, <s>and prediction</s>
1. Training-validation: 2019-2022
2. Test: 2023
3. <strike>prediction: 2024 or last 3 months</strike> (data taken separated on new file)

group_by_year = df.groupby('year')
df_training_validation = pd.concat([group_by_year.get_group(year) for year in range (2019, 2023)])
df_test = group_by_year.get_group(2023)

df_training_validation.to_csv('data/training_validation_mod21c3.csv', index=False)
df_test.to_csv('data/test_mod21c3.csv', index=False)
# df_predict.to_csv('data/predict_mod21c3.csv', index=False)
