==> 4. modeling.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import tensorflow as tf

df = pd.read_csv('data/training_validation_mod21c3.csv')
df.head()

df = df.drop(columns=['year', 'date'])

df.head()

# Data Splitting

# Set the split ratio
split_ratio = 0.9
split_point = int(len(df) * split_ratio)

# Split the data
train_df = df.iloc[:split_point]
validation_df = df.iloc[split_point:]

# Sliding Window
def create_window(dataframe, window_size, slide_size):
    input_window_list = []
    label_window_list = []
    # Iterate over the columns
    for col in tqdm(dataframe.columns):
        # Iterate over the rows in the current column with a sliding window
        for row in range(0, len(dataframe[col]) - (window_size + slide_size) + 1):
            window_range = row + window_size
            
            input_window = dataframe[col].iloc[row:window_range]
            input_window_list.append(input_window)
            
            label_window = dataframe[col].iloc[window_range:window_range + slide_size]
            label_window_list.append(label_window)
    return input_window_list, label_window_list

window_size = 3
slide_size = 1

train_input_window, train_label_window = create_window(train_df, window_size, slide_size)
validation_input_window, validation_label_window = create_window(validation_df, window_size, slide_size)

# Save Train and Validation Data

# with open('data/train_input_window.pkl', 'wb') as f:
#     pickle.dump(train_input_window, f)
#     
# with open('data/train_label_window.pkl', 'wb') as f:
#     pickle.dump(train_label_window, f)
#     
# with open('data/validation_input_window.pkl', 'wb') as f:
#     pickle.dump(validation_input_window, f)
#     
# with open('data/validation_label_window.pkl', 'wb') as f:
#     pickle.dump(validation_label_window, f)

# x_train, y_train = create_window(train_df, window_size, slide_size)
# x_val, y_val = create_window(validation_df, window_size, slide_size)

# Convert to Numpy Array and Reshape

x_train = np.array(train_input_window)
y_train = np.array(train_label_window)
x_val = np.array(validation_input_window)
y_val = np.array(validation_label_window)

x_train = x_train.reshape(x_train.shape[0], x_train.shape[1], 1)
x_val = x_val.reshape(x_val.shape[0], x_val.shape[1], 1)

x_train.shape[1:]

model = tf.keras.models.Sequential([
    tf.keras.layers.LSTM(4, input_shape=(x_train.shape[1:]), return_sequences=True),
    tf.keras.layers.LSTM(4, return_sequences=False),
    tf.keras.layers.Dense(16, activation='relu'),
    tf.keras.layers.Dense(1, activation='relu')
])

model.compile(optimizer='adam', loss='mae', metrics=[tf.keras.metrics.RootMeanSquaredError()])
model.summary()

history = model.fit(x_train, y_train, epochs=10, validation_data=(x_val, y_val))

# TODO: plot loss and metrics

import matplotlib.pyplot as plt

# Plot training & validation loss values
plt.figure(figsize=(12, 6))
plt.plot(history.history['loss'])
plt.plot(history.history['val_loss'])
plt.title('Model loss')
plt.ylabel('Loss')
plt.xlabel('Epoch')
plt.legend(['Train', 'Validation'], loc='upper right')
plt.show()

# Plot training & validation RMSE values
plt.figure(figsize=(12, 6))
plt.plot(history.history['root_mean_squared_error'])
plt.plot(history.history['val_root_mean_squared_error'])
plt.title('Model Root Mean Squared Error')
plt.ylabel('RMSE')
plt.xlabel('Epoch')
plt.legend(['Train', 'Validation'], loc='upper right')
plt.show()

# save model
model.save('models/model_2_3.keras')

# TODO: try with different model configuration
