==> 6. predict.ipynb <==
import ast
import folium
import pandas as pd
import tensorflow as tf
from branca.colormap import LinearColormap

df = pd.read_csv('data/predict_mod21c3.csv')

df.head()

header = pd.concat([df.pop(item) for item in ['year', 'date']], axis=1)

df.head()

header

index_list = list(df.columns.values)
index_list = [ast.literal_eval(string) for string in index_list]

index_list

pred_df = pd.DataFrame(index_list, columns=['longitude', 'latitude'])

pred_df.head()

model = tf.keras.models.load_model('models/model_1_3.keras')

x_pred = df.transpose().values

y_pred = model.predict(x_pred)

y_pred

pred_df['surface_temp'] = y_pred

pred_df.head()

# TODO: convert surface_temp to celsius
pred_df['surface_temp'] = pred_df['surface_temp'] - 273.15

map_center = [pred_df['latitude'].mean(), pred_df['longitude'].mean()]
m = folium.Map(location=map_center, zoom_start=9)
# folium.TileLayer('openstreetmap').add_to(map)

# Define min and max temperature values
min_temp = pred_df['surface_temp'].min()
max_temp = pred_df['surface_temp'].max()

# Define colormap based on min and max temperature values
colormap = LinearColormap(['lightgreen', 'yellow', 'red'], vmin=min_temp, vmax=max_temp)

for index, row in pred_df.iterrows():
    surface_temp = row['surface_temp']
    color = colormap(surface_temp)

    folium.Circle([row['latitude'], row['longitude']],
                  radius=10,
                  color=color,
                  weight=1,
                  fill=True,
                  fill_color=color,
                  fill_opacity=0.4).add_to(m)

colormap.add_to(m)
m.save("map_predict.html")

# display(map)

# TODO: filter dataframe to show on map (surface temp > 35 celsius)
pred_df_filtered = pred_df[pred_df['surface_temp'] > 35]

map_center = [pred_df_filtered['latitude'].mean(), pred_df_filtered['longitude'].mean()]
m = folium.Map(location=map_center, zoom_start=9)

for index, row in pred_df_filtered.iterrows():
    folium.Circle([row['latitude'], row['longitude']],
                  radius=10,
                  color='red',  # fixed color for all markers
                  weight=1,
                  fill=True,
                  fill_color='red',  # fixed fill color for all markers
                  fill_opacity=0.4).add_to(m)

m.save("map_predict_filtered.html")
